refactor(stella-cli): split wrapper_plugin's tests into their own module
The child-turn plane pushed `wrapper_plugin.rs` to 1515 lines, over the
1500-line ratchet and not grandfathered, which is the one remedy the guard
names: split, never a baseline entry.

`mod tests` moves to `wrapper_plugin/tests.rs` verbatim — this crate's
established shape for the same problem (`subagent/tests.rs`,
`plugin_cmd/tests.rs`). No test text changed; 708 + 810 lines.
